ROCm linter and shellcheck warning cleanup (#3715)
Summary:
Pull Request resolved: https://github.com/facebookresearch/faiss/pull/3715

Cleaning up linter and shell warnings after importing the PR. This is done as a follow-up as it's easier to address these in a separate PR since the original PR was authored by an external contributor.

Differential Revision: D60639835

+++ faiss/gpu/impl/InterleavedCodes.cpp
@@ -168,7 +168,7 @@ void unpackInterleavedWord(
         int dims,
         int bitsPerCode) {
     int warpSize = getWarpSizeCurrentDevice();
-    int wordsPerDimBlock = warpSize * bitsPerCode / (8 * sizeof(T));
+    int wordsPerDimBlock = (size_t)warpSize * bitsPerCode / (8 * sizeof(T));
     int wordsPerBlock = wordsPerDimBlock * dims;
     int numBlocks = utils::divUp(numVecs, warpSize);
 
@@ -446,7 +446,7 @@ void packInterleavedWord(
         int dims,
         int bitsPerCode) {
     int warpSize = getWarpSizeCurrentDevice();
-    int wordsPerDimBlock = warpSize * bitsPerCode / (8 * sizeof(T));
+    int wordsPerDimBlock = (size_t)warpSize * bitsPerCode / (8 * sizeof(T));
     int wordsPerBlock = wordsPerDimBlock * dims;
     int numBlocks = utils::divUp(numVecs, warpSize);
